+++ src/KillEmAll/Assets/Scripts/ECS/Raycast/Component/RayCast.cs
@@ -4,7 +4,7 @@ namespace ECS.Raycast.Component
 {
     public struct RayCast : IComponentData
     {
-        public float Length;
+        public int Length;
         public uint CollidesWith;
         public uint BelongsTo;
         public int GroupIndex;
